Add new files and structure to minimalize main.go
