+++ app/main.go
@@ -1,32 +1,19 @@
 package main
 
 import (
-	"encoding/json"
 	"net/http"
 
 	"augustin/database"
+	"augustin/handlers"
 
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/chi/v5/middleware"
 	log "github.com/sirupsen/logrus"
 )
 
-type setting struct {
-	Color string  `json:"color"`
-	Logo  string  `json:"logo"`
-	Price float64 `json:"price"`
-}
-
-type vendor struct {
-	Credit   float64 `json:"credit"`
-	QRcode   string  `json:"qrcode"`
-	IDnumber string  `json:"id-number"`
-}
-
 func initLog() {
 	customFormatter := new(log.TextFormatter)
 	customFormatter.TimestampFormat = "2006-01-02 15:04:05"
-	// customFormatter.FullTimestamp = true
 	log.SetFormatter(customFormatter)
 }
 func main() {
@@ -37,35 +24,6 @@ func main() {
 	s.MountHandlers()
 	log.Info("Server started on port 3000")
 	http.ListenAndServe(":3000", s.Router)
-}
-
-// HelloWorld api Handler
-func HelloWorld(w http.ResponseWriter, r *http.Request) {
-	greeting, err := database.Db.GetHelloWorld()
-	if err != nil {
-		log.Error("QueryRow failed: %v\n", err)
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-	w.Write([]byte(greeting))
-}
-
-func Setting(w http.ResponseWriter, r *http.Request) {
-	marshal_struct, err := json.Marshal(setting{Color: "red", Logo: "/img/Augustin-Logo-Rechteck.jpg", Price: 3.14})
-	if err != nil {
-		log.Info(err)
-		return
-	}
-	w.Write([]byte(marshal_struct))
-}
-
-func Vendor(w http.ResponseWriter, r *http.Request) {
-	marshal_struct, err := json.Marshal(vendor{Credit: 1.61, QRcode: "/img/Augustin-QR-Code.png", IDnumber: "123456789"})
-	if err != nil {
-		log.Info(err)
-		return
-	}
-	w.Write([]byte(marshal_struct))
 }
 
 type Server struct {
@@ -84,11 +42,11 @@ func (s *Server) MountHandlers() {
 	s.Router.Use(middleware.Logger)
 
 	// Mount all handlers here
-	s.Router.Get("/hello", HelloWorld)
+	s.Router.Get("/hello", handlers.HelloWorld)
 
-	s.Router.Get("/settings", Setting)
+	s.Router.Get("/settings", handlers.Setting)
 
-	s.Router.Get("/vendor", Vendor)
+	s.Router.Get("/vendor", handlers.Vendor)
 
 	fs := http.FileServer(http.Dir("img"))
 	s.Router.Handle("/img/*", http.StripPrefix("/img/", fs))
